# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"loginName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"provider": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"app",
									"google",
									"facebook",
									"microsoft"
								]
							}
						]
					},
					"email": {
						"type": "string"
					},
					"mobileNo": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"loginName",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"loginName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"provider": {
						"type": "string",
						"enum": [
							"app",
							"google",
							"facebook",
							"microsoft"
						]
					},
					"email": {
						"type": "string"
					},
					"mobileNo": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"loginName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNo": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChangePasswordInput": {
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MeterDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"accountNo": {
						"type": "string"
					},
					"accountType": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"Prepaid",
									"Postpaid"
								]
							}
						],
						"nullable": true
					},
					"provider": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"enum": [
									"NESCO",
									"DESCO"
								]
							}
						]
					},
					"userId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"accountNo",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMeterInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"accountNo": {
						"type": "string"
					},
					"accountType": {
						"type": "string",
						"enum": [
							"Prepaid",
							"Postpaid"
						]
					},
					"provider": {
						"type": "string",
						"enum": [
							"NESCO",
							"DESCO"
						]
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"accountNo",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMeterInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"accountNo": {
						"type": "string"
					},
					"accountType": {
						"type": "string",
						"enum": [
							"Prepaid",
							"Postpaid"
						]
					},
					"provider": {
						"type": "string",
						"enum": [
							"NESCO",
							"DESCO"
						]
					},
					"userId": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"MeResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"loginName": {
						"type": "string"
					},
					"provider": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"mobileNo": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"loginName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_CreateMeterInput.Exclude_keyofCreateMeterInput.userId__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"accountNo": {
						"type": "string"
					},
					"accountType": {
						"type": "string",
						"enum": [
							"Prepaid",
							"Postpaid"
						]
					},
					"provider": {
						"type": "string",
						"enum": [
							"NESCO",
							"DESCO"
						]
					}
				},
				"required": [
					"name",
					"accountNo"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_CreateMeterInput.userId_": {
				"$ref": "#/components/schemas/Pick_CreateMeterInput.Exclude_keyofCreateMeterInput.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CreateMyMeterInput": {
				"$ref": "#/components/schemas/Omit_CreateMeterInput.userId_"
			},
			"Pick_UpdateMeterInput.Exclude_keyofUpdateMeterInput.userId__": {
				"properties": {
					"name": {
						"type": "string"
					},
					"accountNo": {
						"type": "string"
					},
					"accountType": {
						"type": "string",
						"enum": [
							"Prepaid",
							"Postpaid"
						]
					},
					"provider": {
						"type": "string",
						"enum": [
							"NESCO",
							"DESCO"
						]
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_UpdateMeterInput.userId_": {
				"$ref": "#/components/schemas/Pick_UpdateMeterInput.Exclude_keyofUpdateMeterInput.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"UpdateMyMeterInput": {
				"$ref": "#/components/schemas/Omit_UpdateMeterInput.userId_"
			},
			"ICustomerInfo": {
				"properties": {
					"name": {
						"type": "string"
					},
					"careOf": {
						"type": "string"
					},
					"consumerNo": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"concernOffice": {
						"type": "string"
					},
					"feederName": {
						"type": "string"
					},
					"meterNo": {
						"type": "string"
					},
					"meterType": {
						"type": "string"
					},
					"meterStatus": {
						"type": "string"
					},
					"loadKw": {
						"type": "number",
						"format": "double"
					},
					"tariff": {
						"type": "string"
					},
					"balance": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ICustomerBill": {
				"properties": {
					"billNo": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"month": {
						"type": "string"
					},
					"totalBill": {
						"type": "number",
						"format": "double"
					},
					"lateFee": {
						"type": "number",
						"format": "double"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"paidAmount": {
						"type": "number",
						"format": "double"
					},
					"paymentDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"paymentMethod": {
						"type": "string"
					},
					"paymentStatus": {
						"type": "string"
					},
					"detailUrl": {
						"type": "string"
					},
					"consumerNo": {
						"type": "string"
					}
				},
				"required": [
					"billNo",
					"year",
					"month",
					"totalBill",
					"lateFee",
					"dueDate",
					"paidAmount",
					"paymentDate",
					"paymentMethod",
					"paymentStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICustomerBillResponse": {
				"properties": {
					"customerInfo": {
						"$ref": "#/components/schemas/ICustomerInfo"
					},
					"bills": {
						"items": {
							"$ref": "#/components/schemas/ICustomerBill"
						},
						"type": "array"
					}
				},
				"required": [
					"customerInfo",
					"bills"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRechargeHistory": {
				"properties": {
					"slNo": {
						"type": "number",
						"format": "double"
					},
					"seqNo": {
						"type": "string"
					},
					"tokenNumber": {
						"type": "string"
					},
					"meterRent": {
						"type": "number",
						"format": "double"
					},
					"demandCharge": {
						"type": "number",
						"format": "double"
					},
					"pfcCharge": {
						"type": "number",
						"format": "double"
					},
					"vat": {
						"type": "number",
						"format": "double"
					},
					"paidDebt": {
						"type": "number",
						"format": "double"
					},
					"rebate": {
						"type": "number",
						"format": "double"
					},
					"energyAmount": {
						"type": "number",
						"format": "double"
					},
					"rechargeAmount": {
						"type": "number",
						"format": "double"
					},
					"estimatedUnit": {
						"type": "number",
						"format": "double"
					},
					"rechargeMedia": {
						"type": "string"
					},
					"rechargeDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"remoteRechargeStatus": {
						"type": "string"
					}
				},
				"required": [
					"slNo",
					"seqNo",
					"tokenNumber",
					"meterRent",
					"demandCharge",
					"pfcCharge",
					"vat",
					"paidDebt",
					"rebate",
					"energyAmount",
					"rechargeAmount",
					"estimatedUnit",
					"rechargeMedia",
					"rechargeDate",
					"remoteRechargeStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRechargeHistoryResponse": {
				"properties": {
					"customerInfo": {
						"$ref": "#/components/schemas/ICustomerInfo"
					},
					"rechargeHistories": {
						"items": {
							"$ref": "#/components/schemas/IRechargeHistory"
						},
						"type": "array"
					}
				},
				"required": [
					"customerInfo",
					"rechargeHistories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMonthlyConsumption": {
				"properties": {
					"year": {
						"type": "number",
						"format": "double"
					},
					"month": {
						"type": "string"
					},
					"totalRecharge": {
						"type": "number",
						"format": "double"
					},
					"rebate": {
						"type": "number",
						"format": "double"
					},
					"energyUsage": {
						"type": "number",
						"format": "double"
					},
					"meterRent": {
						"type": "number",
						"format": "double"
					},
					"demandCharge": {
						"type": "number",
						"format": "double"
					},
					"pfcCharge": {
						"type": "number",
						"format": "double"
					},
					"paidDebt": {
						"type": "number",
						"format": "double"
					},
					"vat": {
						"type": "number",
						"format": "double"
					},
					"totalUsageDeduction": {
						"type": "number",
						"format": "double"
					},
					"monthEndBalance": {
						"type": "number",
						"format": "double"
					},
					"energyUsageUnit": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IMonthlyConsumptionResponse": {
				"properties": {
					"customerInfo": {
						"$ref": "#/components/schemas/ICustomerInfo"
					},
					"monthlyConsumptions": {
						"items": {
							"$ref": "#/components/schemas/IMonthlyConsumption"
						},
						"type": "array"
					}
				},
				"required": [
					"customerInfo",
					"monthlyConsumptions"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "nesco-api",
		"version": "1.0.0",
		"description": "NESCO Billing API",
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserInput"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/change-password": {
			"patch": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserDto"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordInput"
							}
						}
					}
				}
			}
		},
		"/meters": {
			"get": {
				"operationId": "GetMeters",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MeterDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Meters"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Meters"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMeterInput"
							}
						}
					}
				}
			}
		},
		"/meters/{id}": {
			"get": {
				"operationId": "GetMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Meters"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Meters"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMeterInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Meters"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Me"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/me/meters": {
			"get": {
				"operationId": "GetMyMeters",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MeterDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Me"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMyMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Me"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMyMeterInput"
							}
						}
					}
				}
			}
		},
		"/me/meters/{id}": {
			"patch": {
				"operationId": "UpdateMyMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Me"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMyMeterInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMyMeter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MeterDto"
								}
							}
						}
					}
				},
				"tags": [
					"Me"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/customers/{customerCode}/bills": {
			"get": {
				"operationId": "GetCustomerBills",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICustomerBillResponse"
								}
							}
						}
					}
				},
				"description": "Get monthly bills of postpaid customers",
				"tags": [
					"Customers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/customers/{customerCode}/recharge-histories": {
			"get": {
				"operationId": "GetCustomerRechargeHistories",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRechargeHistoryResponse"
								}
							}
						}
					}
				},
				"description": "Get recharge histories of prepaid customers",
				"tags": [
					"Customers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/customers/{customerCode}/monthly-consumption": {
			"get": {
				"operationId": "GetCustomerMonthlyConsumption",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMonthlyConsumptionResponse"
								}
							}
						}
					}
				},
				"description": "Get monthly consumption data of prepaid customers",
				"tags": [
					"Customers"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "customerCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}